+++ src/main/java/com/aptner/v3/board/common/reaction/service/CommentReactionService.java
@@ -1,7 +1,7 @@
 package com.aptner.v3.board.common.reaction.service;
 
-import com.aptner.v3.board.comment.domain.Comment;
-import com.aptner.v3.board.comment.service.CommentService;
+import com.aptner.v3.board.comment.Comment;
+import com.aptner.v3.board.comment.CommentService;
 import com.aptner.v3.board.common.reaction.ReactionRepository;
 import com.aptner.v3.board.common.reaction.domain.CommentReaction;
 import org.springframework.stereotype.Service;
