+++ src/main/java/com/aptner/v3/board/common_post/dto/CommonPostDto.java
@@ -1,20 +1,15 @@
 package com.aptner.v3.board.common_post.dto;
 
-import com.aptner.v3.board.comment.dto.CommentDto;
+import com.aptner.v3.board.comment.CommentDto;
 import com.aptner.v3.board.common_post.domain.CommonPost;
 import com.aptner.v3.global.util.ModelMapperUtil;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import org.hibernate.annotations.ColumnDefault;
 import org.modelmapper.ModelMapper;
-import org.modelmapper.config.Configuration;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class CommonPostDto {
